added support for WebDAV REPORT method
Added Sardine.report(), which composes and executes a WebDAV REPORT
request on the passed-in report object.  These report objects must
implement com.github.sardine.report.SardineReport, an interface which
defines a marshal-to-JAXB-object method and a unmarshal-from-Multistatus
method.  (The unmarshal method returns a report-specific object that
we declared as the parameterized generic type when we implemented
SardineReport.)

One report (SyncCollectionReport) is declared in webdav.xsd and in the
report package, but it should be straightforward to add more.