Improve JavaDocs regarding handling of directories
Some operations require the source URL to contain a trailing slash when
referring to a directory, namely move, copy and delete. The put method
on the other hand requires the source URL not pointing to a directory.
In both cases a 301 Moved Permanently status might be issued otherwise
(cf. issue #189).